fix(security): harden 11 security issues from audit (#41 #49 #50 #51 #52 #53 #54 #60 #61 #68 #69) (#76)
fix(security): comprehensive security
  hardening across platform and portal

  Full security audit remediation covering 9
  confirmed vulnerabilities,
  HMAC infrastructure hardening, payment model
   consolidation, and
  35 tasks across 7 workstreams. 73 files
  changed with 60+ regression tests.

  ## Authentication & Token Management

 - Token revocation: switch from POST to     
  DELETE, use request.auth                    
    (self-revocation pattern) instead of      
  body-supplied token key                     
  - Account lockout: wire is_account_locked() 
  and increment_failed_login_attempts()
    into obtain_token and portal_login_api
  with atomic F() counter
  - Uniform 401 responses across all auth   
  failure modes (wrong password,
    locked, inactive, nonexistent) to prevent 
  user enumeration   
  - Mask email PII in auth failure logs via
  _mask_email(); sanitize log
    injection control characters
  - Add token_info endpoint (GET
  /api/users/token/me/) for token-only
    consumers that cannot use HMAC headers
  - ADR-0030: document token auth strategy and
   8 known gaps

## HMAC Infrastructure                      
                                              
  - Replace float timestamps with int across  
  both HMAC systems                           
  - Reject future timestamps (0 <= delta <=   
  window) for preplay prevention
  - Add per-process replay dedup via
  cache.add() with nonce TTL buffer
  - Signature format validation (64-char hex
  pre-filter)
  - Normalize exempt paths with
  _is_auth_exempt() to handle APPEND_SLASH
  - Remove legacy PortalServiceAuthMiddleware
  (weak shared-secret, no replay)
  - Remove redundant body timestamp
  cross-check (body_hash covers integrity)
  - Move rate limiting after signature
  validation (key on verified portal_id)
  - Allow 2s forward clock skew
  (HMAC_NTP_SKEW_SECONDS) for NTP jitter
  - Startup validation for
  PLATFORM_TO_PORTAL_WEBHOOK_SECRET in prod
  settings

- Wire PortalServiceHMACMiddleware into prod
   and staging middleware stacks              
                                              
  ## IP Extraction & Rate Limiting            
                                              
  - Rightmost-trusted-hop XFF model in both   
  services to prevent spoofing                
  - Replace all direct REMOTE_ADDR access with
   get_safe_client_ip()
    (audit/signals, api/customers, api/users,
  common/decorators)
  - Portal: new request_ip.py using
  django-ipware with TRUSTED_PROXY_LIST
    and Cloudflare CF-Ray guard

- Replace
  SessionValidationThrottle(BaseThrottle) with
   ScopedRateThrottle
    to eliminate non-atomic TOCTOU cache
  pattern
  - Atomic cache.add()/cache.incr() counters
  in portal rate limiting
  - API rate limiter fails closed (503) on
  cache errors
  - Rename TRUSTED_PROXY_LIST to
  IPWARE_TRUSTED_PROXY_LIST with Django
    system check portal.W001 for empty proxy
  list
  - Unify RATELIMIT_ENABLE/RATELIMIT_ENABLED
  naming across all settings


## Payment & Billing                        
                                              
  - Payment.confirm_payment(): add
  select_for_update + unique constraint
    on gateway_txn_id + Order row lock for
  double-payment prevention
  - Payment.updated_at field (migration 0018)
  - TERMINAL_PAYMENT_STATUSES frozenset and
  apply_gateway_event() method
  - Refactor StripeWebhookProcessor to use
  canonical payment constants
  - Delete legacy handle_webhook_payment/_hand
  le_stripe_payment_intent
  - Migration 0017: RunPython to convert
  gateway_txn_id="" to NULL before
    applying unique constraint (PostgreSQL
  NULL distinctness)
  - Wire _send_portal_webhook with HMAC-SHA256
   signing (ts + "." + body)

## Nginx & Infrastructure                   
                                              
  - Repeat all 8 security headers in
  static/media location blocks for both
    nginx.conf and nginx-ssl.conf (nginx drops
   server-level add_header
    when any add_header appears in a location
  block)
  - Terraform: firewall_ssh_sources and
  firewall_webmin_sources default=[]
    with validation blocks across all 4
  provider modules

## Security Scanner

  - 9 scanner patterns including REMOTE_ADDR
  access, env-var bypass,
    XFF handling, email-in-logs detection
  - Rewrite _matches_path_glob with
  PurePosixPath.full_match() (Python 3.13)
  - Fix XFF regex and email-in-logs false
  positives

## Semgrep & Static Analysis

  - Resolve all 11 Semgrep findings:
  open-redirect in rate_limiting.py,
    nosemgrep annotations for Django
  Template.render(), innerHTML with
    escapeHtml(), auth log messages,
  blocktrans with asvar

## Testing (60+ new tests)

  - 11 platform unit tests
  (test_api_users_security.py)
  - 8 portal unit tests
  (test_portal_security.py)
  - 14 structural integration tests
  (test_security_hardening.py)
  - 8 webhook hardening tests (body tampering,
   future timestamps, replay)
  - 3 token_info endpoint tests
  - Multi-hop XFF spoofing coverage
  - Byte-identical 401 response verification
  across all failure modes

  <claudiu@captainpragmatic.com>

+++ services/platform/tests/api/test_customer_details_api.py
@@ -86,12 +86,12 @@ class CustomerDetailsAPITests(TestCase):
     )
     def test_customer_details_success_hmac_and_membership(self):
         path = '/api/customers/details/'
-        ts = str(time.time())
+        ts = str(int(time.time()))
         body_dict = {
             'customer_id': self.customer.id,
             'user_id': self.user.id,
             'action': 'get_customer_details',
-            'timestamp': float(ts),
+            'timestamp': int(ts),
         }
         body = json.dumps(body_dict).encode()
         headers = self._headers('POST', path, body, nonce='nonce-success-1', timestamp=ts)
@@ -125,12 +125,12 @@ class CustomerDetailsAPITests(TestCase):
     )
     def test_customer_details_denied_without_membership(self):
         path = '/api/customers/details/'
-        ts = str(time.time())
+        ts = str(int(time.time()))
         body_dict = {
             'customer_id': self.customer.id,
             'user_id': self.other_user.id,  # No membership for this user
             'action': 'get_customer_details',
-            'timestamp': float(ts),
+            'timestamp': int(ts),
         }
         body = json.dumps(body_dict).encode()
         headers = self._headers('POST', path, body, nonce='nonce-nomember-1', timestamp=ts)
@@ -159,12 +159,12 @@ class CustomerDetailsAPITests(TestCase):
     )
     def test_customer_details_invalid_signature_rejected(self):
         path = '/api/customers/details/'
-        ts = str(time.time())
+        ts = str(int(time.time()))
         body_dict = {
             'customer_id': self.customer.id,
             'user_id': self.user.id,
             'action': 'get_customer_details',
-            'timestamp': float(ts),
+            'timestamp': int(ts),
         }
         body = json.dumps(body_dict).encode()
